The question-search process no longer gets stuck on a single level or series of levels.
As more and more questions get asked, the change in difficulty between the last
question and the next question gets smaller and smaller. This has the following
effect:

 - For question banks with low granularity (few levels), the change in difficulty
   would be too small to jump to the next level, resulting in the taker getting stuck
   on a single level.

 - For question banks with high granularity (many levels) and few questions per
   level, the change in difficulty would be too small to jump to the next easier/harder
   level, resulting in the taker getting stuck being asked a series of questions
   that always get easier or always get harder.

In both cases, the inability to alternate between harder and easier questions
caused the measured ability of the taker to skew away from their actual ability
when the test got stuck.

With this change, the questions will continue to bracket the user's measured ability
and will not get stuck above or below the user's ability.