Added function to correctly convert from a linear scale to a logit scale.
The way this was being done before never calculated a logit value when the
difficulty level was at the top or bottom edge of the scale. This gave invalid
ability measures.

As well, this new function caps the logit scale so that answers at the edges of
the scale will not unduly weight the ability measure as the logit value trends
toward infinity.